serialize tags
